refactor(dashboard): split deployment history into Astro shell + focused React islands
Replaces the single monolithic ReactDeploymentHistory island with an
Astro shell (AstroDeploymentHistory.astro) + four small focused
client:only islands. Each island subscribes to exactly the nanostore
atom slice it renders, so a $rows update no longer re-runs the stats
component, and vice versa. Closer to a million.js-style narrow vdom
than the previous single-tree island.

  AstroDeploymentHistory.astro
    Section wrapper + header title + svg + global CSS. SSR-only,
    zero JS for the static chrome.

  ReactDeploymentControls.tsx
    Refresh + live-only filter. Owns the token effect that kicks
    the initial fetch. Subscribes only to $accessToken + $loading
    + $liveOnly.

  ReactDeploymentStats.tsx
    Stats grid (4 cards). Subscribes only to $stats. Renders its
    own grid container so the astro-island wrapper never lands
    inside the parent's grid layout.

  ReactDeploymentRows.tsx
    Owns the full <table> (thead + tbody) so no <astro-island>
    ends up nested inside <tbody> (HTML5 foster-parenting would
    move it out). Row is memo()'d on the row id key so a stats-only
    refresh doesn't re-render historical rows.

  ReactDeploymentError.tsx
    Trivial; renders nothing when $error is null. Avoids paying
    React reconciliation for a missing error banner.

Plus drops the old monolithic ReactDeploymentHistory.tsx and rewires
AstroIDEDashboard.astro to import the new shell. astro-kbve build
clean.
