+++ apps/kbve/astro-kbve/src/components/dashboard/AstroDeploymentHistory.astro
@@ -1,0 +1,196 @@
+---
+import ReactDeploymentControls from './ReactDeploymentControls';
+import ReactDeploymentStats from './ReactDeploymentStats';
+import ReactDeploymentRows from './ReactDeploymentRows';
+import ReactDeploymentError from './ReactDeploymentError';
+---
+
+<section
+	class="deploy-history not-content"
+	aria-label="Firecracker deployment history">
+	<header class="deploy-history-head">
+		<div class="deploy-history-title">
+			<svg
+				xmlns="http://www.w3.org/2000/svg"
+				width="18"
+				height="18"
+				viewBox="0 0 24 24"
+				fill="none"
+				stroke="currentColor"
+				stroke-width="2"
+				stroke-linecap="round"
+				stroke-linejoin="round"
+				aria-hidden="true">
+				<path d="M3 12a9 9 0 1 0 9-9 9.75 9.75 0 0 0-6.74 2.74L3 8">
+				</path>
+				<path d="M3 3v5h5"></path>
+				<path d="M12 7v5l4 2"></path>
+			</svg>
+			<h2>Your Deployments</h2>
+		</div>
+		<ReactDeploymentControls client:only="react" />
+	</header>
+
+	<ReactDeploymentStats client:only="react" />
+
+	<ReactDeploymentError client:only="react" />
+
+	<ReactDeploymentRows client:only="react" />
+</section>
+
+<style is:global>
+	.deploy-history {
+		display: flex;
+		flex-direction: column;
+		gap: 0.75rem;
+		padding: 1rem;
+		border: 1px solid rgba(255, 255, 255, 0.08);
+		border-radius: 0.5rem;
+		background: rgba(255, 255, 255, 0.02);
+		margin-top: 1rem;
+	}
+	.deploy-history-head {
+		display: flex;
+		justify-content: space-between;
+		align-items: center;
+		gap: 0.5rem;
+		flex-wrap: wrap;
+	}
+	.deploy-history-title {
+		display: inline-flex;
+		align-items: center;
+		gap: 0.5rem;
+	}
+	.deploy-history-title h2 {
+		margin: 0;
+		font-size: 1rem;
+	}
+	.deploy-history-controls {
+		display: inline-flex;
+		align-items: center;
+		gap: 0.5rem;
+	}
+	.deploy-history-filter {
+		display: inline-flex;
+		align-items: center;
+		gap: 0.35rem;
+		font-size: 0.78rem;
+		opacity: 0.85;
+		cursor: pointer;
+	}
+	.deploy-history-refresh {
+		display: inline-flex;
+		align-items: center;
+		gap: 0.35rem;
+		padding: 0.35rem 0.6rem;
+		border: 1px solid rgba(255, 255, 255, 0.12);
+		border-radius: 0.25rem;
+		background: rgba(255, 255, 255, 0.04);
+		color: inherit;
+		cursor: pointer;
+		font: inherit;
+		font-size: 0.8rem;
+	}
+	.deploy-history-refresh:disabled {
+		opacity: 0.5;
+		cursor: not-allowed;
+	}
+	.deploy-history-stats {
+		display: grid;
+		grid-template-columns: repeat(auto-fit, minmax(140px, 1fr));
+		gap: 0.5rem;
+	}
+	.deploy-history-stats .stat {
+		padding: 0.6rem 0.75rem;
+		background: rgba(255, 255, 255, 0.03);
+		border: 1px solid rgba(255, 255, 255, 0.06);
+		border-radius: 0.3rem;
+	}
+	.deploy-history-stats .stat-label {
+		font-size: 0.7rem;
+		text-transform: uppercase;
+		letter-spacing: 0.05em;
+		opacity: 0.55;
+	}
+	.deploy-history-stats .stat-value {
+		font-size: 1.15rem;
+		font-family: ui-monospace, SFMono-Regular, Menlo, monospace;
+		margin-top: 0.15rem;
+	}
+	.deploy-history-stats .stat-value.small {
+		font-size: 0.78rem;
+		opacity: 0.85;
+	}
+	.deploy-history-error {
+		padding: 0.5rem 0.7rem;
+		background: rgba(239, 68, 68, 0.12);
+		border: 1px solid rgba(239, 68, 68, 0.4);
+		border-radius: 0.25rem;
+		font-size: 0.8rem;
+	}
+	.deploy-history-table-wrap {
+		overflow-x: auto;
+	}
+	.deploy-history-table {
+		width: 100%;
+		border-collapse: collapse;
+		font-size: 0.78rem;
+	}
+	.deploy-history-table th,
+	.deploy-history-table td {
+		text-align: left;
+		padding: 0.35rem 0.5rem;
+		border-bottom: 1px solid rgba(255, 255, 255, 0.06);
+	}
+	.deploy-history-table th {
+		font-weight: 500;
+		opacity: 0.7;
+	}
+	.deploy-history-table td.mono {
+		font-family: ui-monospace, SFMono-Regular, Menlo, monospace;
+	}
+	.deploy-history-table td.small {
+		font-size: 0.72rem;
+		opacity: 0.85;
+	}
+	.deploy-history-table td.empty {
+		text-align: center;
+		padding: 1.2rem;
+		opacity: 0.55;
+	}
+	.deploy-history-table .badge {
+		display: inline-block;
+		padding: 0.1rem 0.4rem;
+		border-radius: 0.2rem;
+		font-size: 0.7rem;
+		text-transform: uppercase;
+		letter-spacing: 0.04em;
+		font-family: ui-monospace, SFMono-Regular, Menlo, monospace;
+	}
+	.badge-live {
+		background: rgba(34, 197, 94, 0.14);
+		border: 1px solid rgba(34, 197, 94, 0.35);
+		color: rgba(134, 239, 172, 0.95);
+	}
+	.badge-destroyed {
+		background: rgba(255, 255, 255, 0.05);
+		border: 1px solid rgba(255, 255, 255, 0.1);
+		opacity: 0.85;
+	}
+	.badge-unsettled {
+		background: rgba(245, 158, 11, 0.14);
+		border: 1px solid rgba(245, 158, 11, 0.35);
+		color: rgba(252, 211, 77, 0.95);
+	}
+	.spin {
+		animation: deploy-history-spin 1s linear infinite;
+	}
+	@keyframes deploy-history-spin {
+		from {
+			transform: rotate(0deg);
+		}
+		to {
+			transform: rotate(360deg);
+		}
+	}
+</style>
