+++ apps/kbve/astro-kbve/src/components/dashboard/ReactDeploymentError.tsx
@@ -1,0 +1,8 @@
+import { useStore } from '@nanostores/react';
+import { deploymentHistoryService } from './deploymentHistoryService';
+
+export default function ReactDeploymentError() {
+	const error = useStore(deploymentHistoryService.$error);
+	if (!error) return null;
+	return <div className="deploy-history-error">{error}</div>;
+}
